fix(nr-page-table): fix mapping to higher-half (kernel) virtual addresses
nr page table only supports virtual addresses in the lower half
(<= 0x00007fffffffffff), and works incorrectly when the virtual address
is in the higher-half (>= 0xffff800000000000).
When indexing the page table, only the lower 48 bits are used. So this
can be fixed by sending only the lower 48 bits to the nr page table.